+++ src/lib/domain/kjoule_per_day.dart
@@ -1,12 +1,12 @@
 class KJoulePerDay {
   KJoulePerDay({
-    required int kJouleMonday,
-    required int kJouleTuesday,
-    required int kJouleWednesday,
-    required int kJouleThursday,
-    required int kJouleFriday,
-    required int kJouleSaturday,
-    required int kJouleSunday,
+    required double kJouleMonday,
+    required double kJouleTuesday,
+    required double kJouleWednesday,
+    required double kJouleThursday,
+    required double kJouleFriday,
+    required double kJouleSaturday,
+    required double kJouleSunday,
   }) : _kJouleMonday = kJouleMonday,
        _kJouleTuesday = kJouleTuesday,
        _kJouleWednesday = kJouleWednesday,
@@ -15,27 +15,27 @@ class KJoulePerDay {
        _kJouleSaturday = kJouleSaturday,
        _kJouleSunday = kJouleSunday;
 
-  int _kJouleMonday;
-  int _kJouleTuesday;
-  int _kJouleWednesday;
-  int _kJouleThursday;
-  int _kJouleFriday;
-  int _kJouleSaturday;
-  int _kJouleSunday;
+  double _kJouleMonday;
+  double _kJouleTuesday;
+  double _kJouleWednesday;
+  double _kJouleThursday;
+  double _kJouleFriday;
+  double _kJouleSaturday;
+  double _kJouleSunday;
 
-  set kJouleMonday (int kJoule) => _kJouleMonday = kJoule;
-  set kJouleTuesday (int kJoule) => _kJouleTuesday = kJoule;
-  set kJouleWednesday (int kJoule) => _kJouleWednesday = kJoule;
-  set kJouleThursday (int kJoule) => _kJouleThursday = kJoule;
-  set kJouleFriday (int kJoule) => _kJouleFriday = kJoule;
-  set kJouleSaturday (int kJoule) => _kJouleSaturday = kJoule;
-  set kJouleSunday (int kJoule) => _kJouleSunday = kJoule;
+  set kJouleMonday (double kJoule) => _kJouleMonday = kJoule;
+  set kJouleTuesday (double kJoule) => _kJouleTuesday = kJoule;
+  set kJouleWednesday (double kJoule) => _kJouleWednesday = kJoule;
+  set kJouleThursday (double kJoule) => _kJouleThursday = kJoule;
+  set kJouleFriday (double kJoule) => _kJouleFriday = kJoule;
+  set kJouleSaturday (double kJoule) => _kJouleSaturday = kJoule;
+  set kJouleSunday (double kJoule) => _kJouleSunday = kJoule;
 
-  int get kJouleMonday => _kJouleMonday;
-  int get kJouleTuesday => _kJouleTuesday;
-  int get kJouleWednesday => _kJouleWednesday;
-  int get kJouleThursday => _kJouleThursday;
-  int get kJouleFriday => _kJouleFriday;
-  int get kJouleSaturday => _kJouleSaturday;
-  int get kJouleSunday => _kJouleSunday;
+  double get kJouleMonday => _kJouleMonday;
+  double get kJouleTuesday => _kJouleTuesday;
+  double get kJouleWednesday => _kJouleWednesday;
+  double get kJouleThursday => _kJouleThursday;
+  double get kJouleFriday => _kJouleFriday;
+  double get kJouleSaturday => _kJouleSaturday;
+  double get kJouleSunday => _kJouleSunday;
 }
